Update __init__.py
Changed the delimited used in the 'spelled_word - '. '.join(word.upper()' line within the 'handle_spell' function from a perion ( . ) to a semicolon ( ; ).  The period caused an issue pronouncing the letter 'M' when the Google Voice engine was used.  Mimic will be adding the semicolon as a delay as well.